Now handles verbs and messages where dinner texts should be.
Verbs are handled ridiculously easy (with surprising correctness) by simpy checking a dinner text after we've shortened it. If it ends with -te -de -kt we'll assume it's a verb like stekt, wokede, kokt and keep the noun coming after it.

Messages from the nice people in the SiT cantinas are now handled as comments, like so: - "God jul fra oss i Hangaren!". The text is made italic and the quotes and comment sign are both added to make it look more like something an employee might say.