Added a clock and a fourth bus line.
The clock consists of a little HTML, a little CoffeeScript and a good bunch of CSS. The AtB logo has been downsized and moved the spot where the clock now resides. The clock fits in nicely with the rest and has kind of a "TRON"-feel to it. Oh, and it's digital, I considered analog clocks but it wouldn't make sense because the bus lines are displayed in digital time.